+++ modules/__tests__/MemoryHistory-test.js
@@ -152,4 +152,19 @@ describe('a memory history', () => {
       TestSequences.ReturnFalseTransitionHook(history, done)
     })
   })
+
+  describe('with a custom confirm', () => {
+    const getUserConfirmation = (_, callback) => callback(true)
+
+    let history
+    beforeEach(() => {
+      history = createHistory({
+        getUserConfirmation
+      })
+    })
+
+    it('replaces the user confirmation from history creation', (done) => {
+      TestSequences.BlockWithCustomConfirm(history, done)
+    })
+  })
 })