+++ modules/__tests__/TestSequences/index.js
@@ -1,6 +1,7 @@
 export BackButtonTransitionHook from './BackButtonTransitionHook'
 export BlockEverything from './BlockEverything'
 export BlockPopWithoutListening from './BlockPopWithoutListening'
+export BlockWithCustomConfirm from './BlockWithCustomConfirm'
 export DenyPush from './DenyPush'
 export DenyGoBack from './DenyGoBack'
 export DenyGoForward from './DenyGoForward'